DataComparePro_v3_complete/DataComparePro_v3_package/edcp/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name            = "DataComparePro"
version         = "2.0.0"
description     = "Enterprise Data Comparison Platform — multi-batch, audit-logged, self-healing"
readme          = "README.md"
requires-python = ">=3.8"
keywords        = ["data comparison", "ETL validation", "enterprise", "batch", "parquet"]
classifiers     = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
]
dependencies    = [
    "pandas>=1.3.0",
    "openpyxl>=3.0.0",
    "PyYAML>=5.4",
    "flask>=3.0.0",
]

[project.optional-dependencies]
parquet = ["pyarrow>=12.0"]
spark   = ["pyspark>=3.0"]
dev     = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
edcp = "edcp.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["edcp*", "data_compare*"]
